third_party/blink/web_tests/platform/generic/fast/js/unshift-multi-expected.txt

Test for regression against Calling unshift passing more than 1 argument causes array corruption. It also tests some other unshift combinations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS unshift1(1) is ["a"]
PASS unshift1(2) is ["a", "a"]
PASS unshift1(4) is ["a", "a", "a", "a"]
PASS unshift2(1) is ["a", "b"]
PASS unshift2(2) is ["a", "b", "a", "b"]
PASS unshift2(4) is ["a", "b", "a", "b", "a", "b", "a", "b"]
PASS unshift2(10) is ["a", "b", "a", "b", "a", "b", "a", "b", "a", "b", "a", "b", "a", "b", "a", "b", "a", "b", "a", "b"]
PASS unshift5(1) is ["a", "b", "c", "d", "e"]
PASS unshift5(2) is ["a", "b", "c", "d", "e", "a", "b", "c", "d", "e"]
PASS unshift5(6) is ["a", "b", "c", "d", "e", "a", "b", "c", "d", "e", "a", "b", "c", "d", "e", "a", "b", "c", "d", "e", "a", "b", "c", "d", "e", "a", "b", "c", "d", "e"]
PASS successfullyParsed is true

TEST COMPLETE
